@@ -2,14 +2,13 @@ name: location_example
 description: Demonstrates how to use the location plugin.
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   location:
     path: ../
-  url_launcher: ^6.0.2
 
 dev_dependencies:
   flutter_driver:
