@@ -4,8 +4,7 @@ version: 4.2.1
 homepage: https://github.com/Lyokone/flutterlocation
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.4"
+  sdk: '>=3.0.0 <4.0.0'
 
 flutter:
   plugin:
@@ -24,12 +23,10 @@ dependencies:
   flutter:
     sdk: flutter
   location_platform_interface: ^2.2.0
-  location_web: ^3.1.0
   meta: ^1.3.0
 
 dev_dependencies:
   async: ^2.5.0
-  build_runner: ^1.11.5
   flutter_test:
     sdk: flutter
   mockito: ^5.0.0
