@@ -43,7 +43,7 @@ class _GetLocationState extends State<GetLocationWidget> {
       children: <Widget>[
         Text(
           'Location: ' + (_error ?? '${_location ?? "unknown"}'),
-          style: Theme.of(context).textTheme.bodyText1,
+          style: Theme.of(context).textTheme.bodyLarge,
         ),
         Row(
           children: <Widget>[
