@@ -36,7 +36,7 @@ class _ServiceEnabledState extends State<ServiceEnabledWidget> {
       crossAxisAlignment: CrossAxisAlignment.start,
       children: <Widget>[
         Text('Service enabled: ${_serviceEnabled ?? "unknown"}',
-            style: Theme.of(context).textTheme.bodyText1),
+            style: Theme.of(context).textTheme.bodyLarge),
         Row(
           children: <Widget>[
             Container(
